@@ -1,126 +1,56 @@
-name: pdf_learner_v2
-description: "PDF 학습 및 관리 애플리케이션"
+name: pdf_learner
+description: "A new Flutter project."
 publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.2.3 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_web_plugins:
-    sdk: flutter
-  intl: ^0.19.0
-  
-  # UI 관련 패키지
   cupertino_icons: ^1.0.2
-  google_fonts: ^6.2.1
-  flutter_adaptive_scaffold: ^0.1.7
-  adaptive_theme: ^3.4.1
-  responsive_framework: ^1.1.1
-  lottie: ^3.3.1
-  
-  # 상태 관리
+  firebase_core: ^2.24.2
+  firebase_auth: ^4.15.3
+  cloud_firestore: ^4.13.6
+  firebase_storage: ^11.5.6
+  firebase_analytics: ^10.7.4
+  firebase_crashlytics: ^3.4.8
+  firebase_performance: ^0.9.3+8
   provider: ^6.1.1
-  
-  # 저장소 관련
   shared_preferences: ^2.2.2
-  shared_preferences_windows: ^2.3.2  # Windows 지원 추가
+  flutter_dotenv: ^5.1.0
+  get_it: ^7.6.4
+  injectable: ^2.3.2
+  freezed_annotation: ^2.4.1
+  json_annotation: ^4.9.0
+  google_fonts: ^6.1.0
+  pdf: ^3.10.7
   path_provider: ^2.1.1
-  flutter_secure_storage: ^9.2.4
-  sqflite: ^2.3.0
-  
-  # Firebase 관련
-  firebase_core: ^3.12.1
-  firebase_auth: ^5.5.1
-  cloud_firestore: ^5.6.5
-  firebase_storage: ^12.4.1
-  firebase_analytics: ^11.4.1
-  firebase_crashlytics: ^4.3.1
-  google_sign_in: ^6.3.0
-  
-  # 플랫폼 인터페이스들
-  firebase_core_platform_interface: ^5.0.0
-  firebase_auth_platform_interface: ^7.0.9
-  cloud_firestore_platform_interface: ^6.0.10
-  firebase_storage_platform_interface: ^5.1.11
-  
-  # PDF 관련
-  syncfusion_flutter_pdfviewer: ^24.2.8
-  syncfusion_flutter_pdf: ^24.1.41
-  flutter_pdfview: ^1.4.0
-  pdf: ^3.10.4
-  printing: ^5.12.0
-  
-  # 파일 선택 관련
   file_picker: ^6.1.1
-  file_selector: ^1.0.3
-  file_selector_windows: ^0.9.3
-  file_selector_macos: ^0.9.4
-  file_selector_linux: ^0.9.3
-  file_selector_android: ^0.5.1
-  file_selector_ios: ^0.5.3
-  cross_file: ^0.3.4
-  desktop_drop: ^0.5.0
-  
-  # 네트워크 및 유틸리티
-  http: ^1.1.0
-  url_launcher: ^6.2.4
+  http: ^1.1.2
+  intl: ^0.19.0
+  flutter_local_notifications: ^16.2.0
+  in_app_purchase: ^3.1.11
+  flutter_secure_storage: ^9.0.0
   js: ^0.6.7
-  platform: ^3.1.3
-  device_info_plus: ^9.1.2
-  package_info_plus: ^4.2.0
-  flutter_dotenv: ^5.2.1
-  universal_html: ^2.2.4
-  crypto: ^3.0.3
-  encrypt: ^5.0.3
-  flutter_svg: ^2.0.9
-  
-  # 윈도우 관련
-  window_size:
-    git:
-      url: https://github.com/google/flutter-desktop-embedding
-      path: plugins/window_size
-  flutter_acrylic: ^1.1.3
-  window_manager: ^0.4.3
-  flutter_riverpod: ^2.5.1
-  uuid: ^4.2.1
-  animate_do: ^4.2.0
-  google_mobile_ads: ^3.1.0
-  pdf_render: ^1.4.2  # PDF 렌더링 패키지 추가
-  flutter_staggered_grid_view: ^0.7.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  integration_test:
-    sdk: flutter
-  mockito: ^5.4.4
+  flutter_lints: ^2.0.0
   build_runner: ^2.4.8
-  flutter_lints: ^5.0.0
-  fake_cloud_firestore: ^3.1.0
+  freezed: ^2.4.7
+  json_serializable: ^6.7.1
+  injectable_generator: ^2.4.1
+  mockito: ^5.4.4
 
 flutter:
   uses-material-design: true
-  
-  # 다국어 지원 설정
-  generate: true
-  
-  # 애셋 설정
   assets:
+    - .env
     - assets/images/
     - assets/icons/
-    - assets/images/google_logo.svg
-    - assets/google_logo.svg
-    - .env
-    
-  # 폰트 설정 (필요시 추가)
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic 
+    - assets/google_logo.svg 
