@@ -3,6 +3,7 @@
 import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
 import 'package:flutter/foundation.dart'
     show defaultTargetPlatform, kIsWeb, TargetPlatform;
+import 'package:flutter_dotenv/flutter_dotenv.dart';
 
 /// 기본 Firebase 구성 옵션
 class DefaultFirebaseOptions {
@@ -29,50 +30,51 @@ class DefaultFirebaseOptions {
         );
       default:
         throw UnsupportedError(
-          'DefaultFirebaseOptions are not supported for this platform.',
+          'DefaultFirebaseOptions have not been configured for $defaultTargetPlatform - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
         );
     }
   }
 
   /// Firebase 웹 앱 설정
-  static const FirebaseOptions web = FirebaseOptions(
-    apiKey: 'AIzaSyB7Zj4PNPF80RL9qB1s7w5k3ZaBKVeYRo4',
-    appId: '1:189136888100:web:e4af7ecbaa8beb37c9c3d1',
-    messagingSenderId: '189136888100',
-    projectId: 'pdf-learner',
-    authDomain: 'pdf-learner.firebaseapp.com',
-    storageBucket: 'pdf-learner.appspot.com',
-    measurementId: 'G-LY7QXCN3ND',
+  static final FirebaseOptions web = FirebaseOptions(
+    apiKey: dotenv.env['FIREBASE_API_KEY'] ?? '',
+    appId: dotenv.env['FIREBASE_APP_ID'] ?? '',
+    messagingSenderId: dotenv.env['FIREBASE_MESSAGING_SENDER_ID'] ?? '',
+    projectId: dotenv.env['FIREBASE_PROJECT_ID'] ?? '',
+    authDomain: dotenv.env['FIREBASE_AUTH_DOMAIN'] ?? '',
+    storageBucket: dotenv.env['FIREBASE_STORAGE_BUCKET'] ?? '',
+    measurementId: dotenv.env['FIREBASE_MEASUREMENT_ID'] ?? '',
   );
 
   /// Firebase Android 앱 설정
-  static const FirebaseOptions android = FirebaseOptions(
-    apiKey: 'AIzaSyDWLZ5nGK5tXm0_FCoCYKOc50rKO_fUZhI',
-    appId: '1:189136888100:android:e8a1b37a9a48d146c9c3d1',
-    messagingSenderId: '189136888100',
-    projectId: 'pdf-learner',
-    storageBucket: 'pdf-learner.appspot.com',
+  static final FirebaseOptions android = FirebaseOptions(
+    apiKey: dotenv.env['FIREBASE_API_KEY'] ?? '',
+    appId: dotenv.env['FIREBASE_APP_ID'] ?? '',
+    messagingSenderId: dotenv.env['FIREBASE_MESSAGING_SENDER_ID'] ?? '',
+    projectId: dotenv.env['FIREBASE_PROJECT_ID'] ?? '',
+    storageBucket: dotenv.env['FIREBASE_STORAGE_BUCKET'] ?? '',
   );
 
   /// Firebase iOS 앱 설정
-  static const FirebaseOptions ios = FirebaseOptions(
-    apiKey: 'AIzaSyBs4XXXXXXXXXXXXXXXXXXXXXXXX',
-    appId: '1:189136888100:ios:0000000000000000000000',
-    messagingSenderId: '189136888100',
-    projectId: 'pdf-learner',
-    storageBucket: 'pdf-learner.appspot.com',
-    iosClientId: 'XXXXX.apps.googleusercontent.com',
-    iosBundleId: 'com.example.pdfLearnerV2',
+  static final FirebaseOptions ios = FirebaseOptions(
+    apiKey: dotenv.env['FIREBASE_API_KEY'] ?? '',
+    appId: dotenv.env['FIREBASE_APP_ID'] ?? '',
+    messagingSenderId: dotenv.env['FIREBASE_MESSAGING_SENDER_ID'] ?? '',
+    projectId: dotenv.env['FIREBASE_PROJECT_ID'] ?? '',
+    storageBucket: dotenv.env['FIREBASE_STORAGE_BUCKET'] ?? '',
+    iosClientId: dotenv.env['FIREBASE_IOS_CLIENT_ID'] ?? '',
+    iosBundleId: dotenv.env['FIREBASE_IOS_BUNDLE_ID'] ?? '',
   );
 
   /// Firebase macOS 앱 설정
-  static const FirebaseOptions macos = FirebaseOptions(
-    apiKey: 'AIzaSyBs4XXXXXXXXXXXXXXXXXXXXXXXX',
-    appId: '1:189136888100:ios:0000000000000000000000',
-    messagingSenderId: '189136888100',
-    projectId: 'pdf-learner',
-    storageBucket: 'pdf-learner.appspot.com',
-    iosClientId: 'XXXXX.apps.googleusercontent.com',
-    iosBundleId: 'com.example.pdfLearnerV2',
+  static final FirebaseOptions macos = FirebaseOptions(
+    apiKey: dotenv.env['FIREBASE_API_KEY'] ?? '',
+    appId: dotenv.env['FIREBASE_APP_ID'] ?? '',
+    messagingSenderId: dotenv.env['FIREBASE_MESSAGING_SENDER_ID'] ?? '',
+    projectId: dotenv.env['FIREBASE_PROJECT_ID'] ?? '',
+    storageBucket: dotenv.env['FIREBASE_STORAGE_BUCKET'] ?? '',
+    iosClientId: dotenv.env['FIREBASE_IOS_CLIENT_ID'] ?? '',
+    iosBundleId: dotenv.env['FIREBASE_IOS_BUNDLE_ID'] ?? '',
   );
 } 
