@@ -1,108 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-import 'package:google_fonts/google_fonts.dart';
+import 'package:shared_preferences.dart';
 
-/// 테마 관리 서비스
-class ThemeService with ChangeNotifier {
+class ThemeService extends ChangeNotifier {
   static const String _themeKey = 'theme_mode';
-  
   final SharedPreferences _prefs;
   ThemeMode _themeMode;
-  
-  ThemeService(this._prefs) : _themeMode = _loadThemeMode(_prefs);
-  
-  /// 현재 테마 모드
+
+  ThemeService(this._prefs) : _themeMode = ThemeMode.values[_prefs.getInt(_themeKey) ?? 0];
+
   ThemeMode get themeMode => _themeMode;
-  
-  /// 현재 테마가 다크 모드인지 여부
-  bool get isDarkMode => _themeMode == ThemeMode.dark;
-  
-  /// 라이트 테마 정의
-  ThemeData get lightTheme {
-    return ThemeData(
-      colorScheme: ColorScheme.fromSeed(
-        seedColor: Colors.blue,
-        brightness: Brightness.light,
-      ),
-      useMaterial3: true,
-      fontFamily: GoogleFonts.notoSans().fontFamily,
-      appBarTheme: const AppBarTheme(
-        elevation: 0,
-        centerTitle: false,
-      ),
-      cardTheme: const CardTheme(
-        elevation: 2,
-        margin: EdgeInsets.symmetric(vertical: 8, horizontal: 16),
-      ),
-      elevatedButtonTheme: ElevatedButtonThemeData(
-        style: ElevatedButton.styleFrom(
-          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(8),
-          ),
-        ),
-      ),
-      inputDecorationTheme: InputDecorationTheme(
-        border: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(8),
-        ),
-        contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
-      ),
-    );
-  }
-  
-  /// 다크 테마 정의
-  ThemeData get darkTheme {
-    return ThemeData(
-      colorScheme: ColorScheme.fromSeed(
-        seedColor: Colors.blue,
-        brightness: Brightness.dark,
-      ),
-      useMaterial3: true,
-      fontFamily: GoogleFonts.notoSans().fontFamily,
-      appBarTheme: const AppBarTheme(
-        elevation: 0,
-        centerTitle: false,
-      ),
-      cardTheme: const CardTheme(
-        elevation: 2,
-        margin: EdgeInsets.symmetric(vertical: 8, horizontal: 16),
-      ),
-      elevatedButtonTheme: ElevatedButtonThemeData(
-        style: ElevatedButton.styleFrom(
-          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(8),
-          ),
-        ),
-      ),
-      inputDecorationTheme: InputDecorationTheme(
-        border: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(8),
-        ),
-        contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
-      ),
-    );
-  }
-  
-  /// 테마 모드 변경
+
   Future<void> setThemeMode(ThemeMode mode) async {
     _themeMode = mode;
     await _prefs.setInt(_themeKey, mode.index);
     notifyListeners();
   }
-  
-  /// 다크 모드 전환
-  Future<void> toggleTheme() async {
-    final newMode = isDarkMode ? ThemeMode.light : ThemeMode.dark;
-    await setThemeMode(newMode);
-  }
-  
-  /// 저장된 테마 모드 불러오기
-  static ThemeMode _loadThemeMode(SharedPreferences prefs) {
-    final index = prefs.getInt(_themeKey);
-    if (index == null) return ThemeMode.system;
-    
-    return ThemeMode.values[index];
-  }
+
+  bool get isDarkMode => _themeMode == ThemeMode.dark;
 } 
