@@ -1,16 +1,19 @@
-// Firebase 스텁 파일
-// 웹이 아닌 환경에서 Firebase 없이 테스트할 수 있도록 함
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_storage/firebase_storage.dart';
+import 'package:firebase_analytics/firebase_analytics.dart';
+import 'package:firebase_crashlytics/firebase_crashlytics.dart';
+import 'package:firebase_performance/firebase_performance.dart';
 
-class Firebase {
-  static Future<void> initializeApp({dynamic options}) async {
-    // Firebase 초기화가 필요 없는 더미 메소드
-    print('Firebase 초기화 스킵 (스텁 사용)');
-    return;
+class StubFirebase {
+  static Future<void> initialize() async {
+    // Firebase 초기화 로직
   }
-}
 
-class FirebaseApp {
-  final String name;
-  
-  FirebaseApp(this.name);
+  static FirebaseAuth get auth => FirebaseAuth.instance;
+  static FirebaseFirestore get firestore => FirebaseFirestore.instance;
+  static FirebaseStorage get storage => FirebaseStorage.instance;
+  static FirebaseAnalytics get analytics => FirebaseAnalytics.instance;
+  static FirebaseCrashlytics get crashlytics => FirebaseCrashlytics.instance;
+  static FirebasePerformance get performance => FirebasePerformance.instance;
 } 
