@@ -11,20 +11,11 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:pdf_learner_v2/main.dart';
 
 void main() {
-  testWidgets('Counter increments smoke test', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
+  testWidgets('기본 앱 테스트', (WidgetTester tester) async {
+    // 앱 빌드
     await tester.pumpWidget(const MyApp());
 
-    // Verify that our counter starts at 0.
-    expect(find.text('0'), findsOneWidget);
-    expect(find.text('1'), findsNothing);
-
-    // Tap the '+' icon and trigger a frame.
-    await tester.tap(find.byIcon(Icons.add));
-    await tester.pump();
-
-    // Verify that our counter has incremented.
-    expect(find.text('0'), findsNothing);
-    expect(find.text('1'), findsOneWidget);
+    // 기본 위젯 존재 확인
+    expect(find.byType(MaterialApp), findsOneWidget);
   });
 }
